requirements.txt
# Core dependencies
python-telegram-bot==13.15
pyyaml==6.0.2
requests==2.32.3
pillow==11.2.1
pydantic==2.11.5
schedule==1.2.2
keyboard==0.13.5

# For image captioning
transformers==4.52.3
torch==2.7.0